+++ src/main/java/lotto/domain/LottoStore.java
@@ -2,6 +2,7 @@ package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.util.UniqueRandomNumberProvider;
 
 public class LottoStore {
