+++ src/main/java/lotto/util/UniqueRandomNumberProvider.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import java.util.Collections;
 import java.util.List;
