+++ src/test/java/lotto/util/UniqueRandomNumberProviderTest.java
@@ -1,11 +1,11 @@
-package lotto.domain;
+package lotto.util;
 
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import lotto.domain.UniqueRandomNumberProvider;
+import lotto.util.UniqueRandomNumberProvider;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
